Added two types and one attribute
Added value_set and value_map types. These allow the specification
of run-time managed state objects at the IR level.

Added allowed_actions as an attribute to tables. This is currently
ignored by AIR and IRI, but allows the specification of the
actions that should be allowed to be associated with entries
in a given table.