Updates to AIR instance for manifest support
Support processing a group of files as a single yaml
chunk of input. A new aggregator class was added to AIR.
When add_content is called, if it is passed a list of
filenames, the contents of all those files is accumulated
and then processed as a single chunk of input.